# .sdd/specs/ai-fast-grading/feat-speaking-ai-grading/contracts/speaking-grading.openapi.yaml
openapi: 3.1.0
info:
  title: Hợp đồng API chấm Speaking của IELTSZone
  version: 0.3.0-ai-estimated
  description: >-
    Hợp đồng triển khai cho luồng Speaking grading bất đồng bộ: grader=ai trả đủ
    bốn tiêu chí từ transcript + audio dưới nhãn AI Estimated Band hoặc failed;
    không tự handoff tutor. Calibration/RFC vẫn là cổng cho mức đã hiệu chuẩn.
servers:
  - url: /api/v1/submissions
security:
  - cookieAuth: []
paths:
...
  /{submissionId}/audio-url:
    get:
      summary: Cấp signed download URL ngắn hạn
      operationId: getSpeakingAudioUrl
      x-allowed-roles: [student, tutor, admin]
      x-authorization-scope: owner-or-assigned-tutor-or-admin
      parameters:
        - name: submissionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          required: true
          schema:
            type: string
            const: speaking
      responses:
        '200':
          description: Signed download URL
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                const: private, no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudioUrlEnvelope'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/ApiError'
        '404':
          $ref: '#/components/responses/ApiError'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
...
  responses:
    ApiError:
      description: Lỗi theo envelope chuẩn
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    AudioUrlData:
      type: object
      additionalProperties: false
      required: [url, expires_at]
      properties:
        url:
          type: string
          format: uri
        expires_at:
          type: string
          format: date-time
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          oneOf:
            - type: object
              additionalProperties: true
            - type: 'null'
    SuccessEnvelopeBase:
      type: object
      additionalProperties: false
      required: [success, data, error, meta]
      properties:
        success:
          const: true
        data: {}
        error:
          type: 'null'
        meta:
          type: object
          additionalProperties: true
...
    AudioUrlEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelopeBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/AudioUrlData'
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [success, data, error, meta]
      properties:
        success:
          const: false
        data:
          type: 'null'
        error:
          $ref: '#/components/schemas/ApiError'
        meta:
          type: object
          additionalProperties: true